requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
flake8-html==0.4.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.6
